Relocate dot-page-container
