+++ src/app/portlets/dot-edit-page/content/services/dot-container-contentlet.service.spec.ts
@@ -1,5 +1,5 @@
 import { DotContainerContentletService } from './dot-container-contentlet.service';
-import { DotPageContainer } from '../../../dot-edit-page/shared/models/dot-page-container.model';
+import { DotPageContainer } from '../../../../shared/models/dot-page-container/dot-page-container.model';
 import { DotPageContent } from '../../../dot-edit-page/shared/models/dot-page-content.model';
 import { DotCMSContentType } from 'dotcms-models';
 import { dotcmsContentTypeBasicMock } from '@tests/dot-content-types.mock';
