+++ src/app/portlets/dot-edit-page/content/services/dot-edit-content-html/dot-edit-content-html.service.spec.ts
@@ -1,31 +1,30 @@
-import { of, Observable } from 'rxjs';
+import { Injectable } from '@angular/core';
+
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed, waitForAsync } from '@angular/core/testing';
+
+import { of, Observable } from 'rxjs';
+import { ConfirmationService } from 'primeng/api';
+
 import { DotEditContentHtmlService, DotContentletAction } from './dot-edit-content-html.service';
 import { DotEditContentToolbarHtmlService } from '../html/dot-edit-content-toolbar-html.service';
 import { DotContainerContentletService } from '../dot-container-contentlet.service';
 import { DotDragDropAPIHtmlService } from '../html/dot-drag-drop-api-html.service';
 import { DotDOMHtmlUtilService } from '../html/dot-dom-html-util.service';
 import { DotMessageService } from '@services/dot-message/dot-messages.service';
-import { MockDotMessageService } from '../../../../../test/dot-message-service.mock';
+import { MockDotMessageService } from '@tests/dot-message-service.mock';
 import { LoggerService, StringUtils, CoreWebService } from 'dotcms-js';
 import { DotAlertConfirmService } from '@services/dot-alert-confirm/dot-alert-confirm.service';
-import { DotPageContent } from '../../../../dot-edit-page/shared/models/dot-page-content.model';
-import {
-    mockDotLayout,
-    mockDotRenderedPage,
-    mockDotPage
-} from '../../../../../test/dot-page-render.mock';
+import { mockDotLayout, mockDotRenderedPage, mockDotPage } from '@tests/dot-page-render.mock';
 import { DotCMSContentType } from 'dotcms-models';
 import { DotLicenseService } from '@services/dot-license/dot-license.service';
-import { Injectable } from '@angular/core';
 import { DotPageRenderState } from '@portlets/dot-edit-page/shared/models/dot-rendered-page-state.model';
-import { mockUser } from '../../../../../test/login-service.mock';
+import { mockUser } from '@tests/login-service.mock';
 import { PageModelChangeEventType } from './models';
 import { dotcmsContentTypeBasicMock } from '@tests/dot-content-types.mock';
-import { DotPageRender, DotPageContainer } from '@portlets/dot-edit-page/shared/models';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { DotPageContent, DotPageRender } from '@portlets/dot-edit-page/shared/models';
 import { CoreWebServiceMock } from '@tests/core-web.service.mock';
-import { ConfirmationService } from 'primeng/api';
+import { DotPageContainer } from '@models/dot-page-container/dot-page-container.model';
 
 @Injectable()
 class MockDotLicenseService {
