+++ src/app/portlets/dot-edit-page/content/services/dot-edit-content-html/models/page-model-change-event.ts
@@ -1,4 +1,4 @@
-import { DotPageContainer } from '@portlets/dot-edit-page/shared/models/dot-page-container.model';
+import { DotPageContainer } from '@models/dot-page-container/dot-page-container.model';
 import { PageModelChangeEventType } from './page-model-change-event.type';
 
 export interface PageModelChangeEvent {
