+++ src/app/shared/models/dot-edit-layout-designer/dot-layout-column.model.ts
@@ -1,4 +1,4 @@
-import { DotPageContainer } from '@portlets/dot-edit-page/shared/models';
+import { DotPageContainer } from '@models/dot-page-container/dot-page-container.model';
 
 export interface DotLayoutColumn {
     containers: DotPageContainer[];
